// src/features/store/types.ts
export interface NavigationItem {
    title: string;
    linkTo: string;
};

export interface MobileDevices {
    name: string;
    diagonal: string;
    memory: string;
    serialNumber: string;
    productNumber: string;
    characteristics: string;
    width: string,
    height: string,
    depth: string,
    weight: string;
    image: string;
}